Initial split to abstract JMX operator
